+++ crates/engine/src/credential/rotation.rs
@@ -4,6 +4,8 @@
 //! this module while canonical contract/state types remain in
 //! `nebula_credential::rotation`.
 
+pub mod scheduler;
+
 pub use nebula_credential::rotation::{
     RotationError, RotationResult,
     blue_green::{BlueGreenRotation, BlueGreenState},
@@ -12,9 +14,9 @@ pub use nebula_credential::rotation::{
         cleanup_expired_credentials, track_credential_usage,
     },
     policy::{BeforeExpiryConfig, ManualConfig, PeriodicConfig, RotationPolicy, ScheduledConfig},
-    scheduler::{ExpiryMonitor, PeriodicScheduler, ScheduledRotation},
     transaction::{
         BackupId, ManualRotation, OptimisticLock, RollbackStrategy, RotationId,
         RotationTransaction, TransactionPhase, ValidationResult,
     },
 };
+pub use scheduler::{ExpiryMonitor, PeriodicScheduler, ScheduledRotation};
